+++ src/controllers/GameHistory.js
@@ -81,3 +81,5 @@ exports.getLastProtectedPlayer = async gameId => {
 };
 
 exports.getLastVotePlay = gameId => this.findOne({ gameId, "play.action": "vote" }, null, { sort: { createdAt: -1 } });
+
+exports.getLastSistersPlay = gameId => this.findOne({ gameId, "play.source": "two-sisters" }, null, { sort: { createdAt: -1 } });