+++ tests/e2e/game/game-creation.test.js
@@ -34,6 +34,14 @@ const playersWithoutVillagers = [
     { name: "Dyg", role: "werewolf" },
     { name: "Deg", role: "werewolf" },
 ];
+const playersWithOnlyOneSister = [
+    { name: "Dig", role: "villager" },
+    { name: "Doug", role: "two-sisters" },
+    { name: "Dag", role: "werewolf" },
+    { name: "Dug", role: "werewolf" },
+    { name: "Dyg", role: "werewolf" },
+    { name: "Deg", role: "werewolf" },
+];
 let token, token2, game, game2, queryString;
 
 describe("A - Game creation", () => {
@@ -130,6 +138,17 @@ describe("A - Game creation", () => {
             .end((err, res) => {
                 expect(res).to.have.status(400);
                 expect(res.body.type).to.equals("BAD_REQUEST");
+                done();
+            });
+    });
+    it("👭 Can't create game a with only one sister (POST /games)", done => {
+        chai.request(app)
+            .post("/games")
+            .set({ Authorization: `Bearer ${token}` })
+            .send({ players: playersWithOnlyOneSister })
+            .end((err, res) => {
+                expect(res).to.have.status(400);
+                expect(res.body.type).to.equals("SISTERS_MUST_BE_TWO");
                 done();
             });
     });